requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf